+++ web/src/components/Client.vue
@@ -206,7 +206,7 @@ export default {
       if(this.graphReady) {
         let graph = this.$refs.graph;
         let value = graph.focusedValues[mKey];
-        if(value) {
+        if(!(value === undefined)) {
           return Number(value).format(this.configMap[mKey].format);
         }
       }